Create SkipTheWhiteBulk-OpenSource
This source is coded by Hongshuai Wang  by looking the video educate on Bilibili, But these were not same as the video due to some change. And I created it in "Windows Desktop Application" on visual studio 2019, so unbelievable is it that I only key some code into their  line.